refactor: rename local tree compatibility layer to #tree-state
