+++ src/components/commandLineCharacters/commandLineCharacters.store.js
@@ -1,4 +1,4 @@
-import { toFlatGroups, toFlatItems } from "#insieme-compat";
+import { toFlatGroups, toFlatItems } from "#tree-state";
 
 export const createInitialState = () => ({
   mode: "current",
