+++ src/components/commandLineSoundEffects/commandLineSoundEffects.handlers.js
@@ -1,4 +1,4 @@
-import { toFlatItems } from "#insieme-compat";
+import { toFlatItems } from "#tree-state";
 import { nanoid } from "nanoid";
 
 export const handleAfterMount = async (deps) => {
