+++ src/components/linesEditor/linesEditor.store.js
@@ -1,4 +1,4 @@
-import { toFlatItems } from "#insieme-compat";
+import { toFlatItems } from "#tree-state";
 
 export const createInitialState = () => ({
   ready: false,
