+++ src/deps/infra/eventStore.js
@@ -4,7 +4,7 @@ import {
   toFlatItems,
   toFlatGroups,
   toTreeStructure,
-} from "#insieme-compat";
+} from "#tree-state";
 
 // Re-export utility functions from insieme
 export { toFlatItems, toFlatGroups, toTreeStructure };
