+++ src/pages/scenes/scenes.store.js
@@ -1,4 +1,4 @@
-import { toFlatGroups, toFlatItems } from "#insieme-compat";
+import { toFlatGroups, toFlatItems } from "#tree-state";
 import { getSectionPresentation } from "../../utils/sectionPresentation.js";
 
 const form = {
