+++ src/pages/tweens/tweens.store.js
@@ -1,4 +1,4 @@
-import { toFlatGroups, toFlatItems } from "#insieme-compat";
+import { toFlatGroups, toFlatItems } from "#tree-state";
 import { resetState } from "./tweens.constants";
 
 const form = {
